Improve performance in calculating framerate
As tested on MBP, although framerate is not higher than before, but this
reduces the calculation operations needed in the future. Notice double
that we define for relevant framerate variables.

+++ src/foundation/common.c
@@ -2,9 +2,9 @@
 
 KRR_WINDOW* gWindow = NULL;
 
-float common_frameTime = 0.0f;
+double common_frameTime = 0.0f;
 #ifndef DISABLE_FPS_CALC
 int common_frameCount = 0;
-float common_frameAccumTime = 0.0f;
-float common_avgFPS = 0.0f;
+double common_frameAccumTime = 0.0f;
+double common_avgFPS = 0.0f;
 #endif
